Add new 2pt fixture
We want to reuse the same test data the two-part tariff annual process-billing-period service is using.
